% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.R
\name{feed}
\alias{feed}
\title{Title}
\usage{
feed(l, parms = livestock$parms, factor = 1)
}
\arguments{
\item{parms}{a list of parameter values. defaults to the livestock default parameter set.}
}
\value{
a single value: the expected number of plants dying from livestock
  feeding on the given landscape l under the given parameter combination.
}
\description{
Title
}

