% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_list.R
\name{init_list}
\alias{init_list}
\title{Initialize a list of random landscape objects}
\usage{
init_list(n, cover = NULL, width = 200, height = width,
  runif_range = c(0, 1))
}
\arguments{
\item{n}{An integer number. Number of landscape objects to generate.}

\item{cover}{A single numeric value. The average vegetation cover of the
landscape. Defaults to NULL, which calls a random number generator drawing
initial cover from within the range given in  \code{runif_range}.}

\item{width}{An integer number. The width of the lattice to generate.
Defaults to 200.}

\item{height}{An integer number. defaults to the same value as width to
produce square landscape objects.}

\item{runif_range}{A numeric vector of length two. Gives the lower and upper
boundary of the random number generator for initial cover.}
}
\value{
A list of n landscape objects. The output object is of class
  "landscape_list", which can be processed by the functions
  \code{list_update()} and \code{summary()}.
}
\description{
Initialize a list of random landscape objects
}
\examples{

L0 <- init_list(100, runif_range = c(0.6,0.99), width = 25)

}
