% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.landscape_list}
\alias{summary.landscape_list}
\title{Summary method for landscape lists.}
\usage{
\method{summary}{landscape_list}(L, parms = livestock$parms)
}
\arguments{
\item{L}{A list of landscape objects (may be of class "landscape_list").}
}
\value{
A list of statistical output containing the following entries: 
\itemize{
 \item{\code{$cover}}{Stuff}
 \item{\code{$local}}{Stuff}
 \item{\code{$clustering}}{Stuff}
 \item{\code{$kernel}}{Stuff}
}
}
\description{
Summary method for landscape lists.
}
\examples{

L0 <- init_list(100, runif_range = c(0.6,0.99), width = 50)
summary(L0)

}

