% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vioplot_betakernel.R
\name{vioplot_beta}
\alias{vioplot_beta}
\title{Title}
\usage{
vioplot_beta(x, ..., range = 1.5, h = NULL, ylim = NULL, names = NULL,
  horizontal = FALSE, col = "white", border = "black", lty = 1,
  lwd = 1, rectCol = "black", colMed = "white", pchMed = 19, at,
  add = FALSE, wex = 1, drawRect = TRUE)
}
\arguments{
\item{x}{}

\item{...}{}

\item{range}{}

\item{h}{}

\item{ylim}{}

\item{names}{}

\item{horizontal}{}

\item{col}{}

\item{border}{}

\item{lty}{}

\item{lwd}{}

\item{rectCol}{}

\item{colMed}{}

\item{pchMed}{}

\item{at}{}

\item{add}{}

\item{wex}{}

\item{drawRect}{}
}
\description{
Title
}
\examples{

vioplot_beta(summary(L100)$feed)
}

