% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betakernel.R
\name{betakernel}
\alias{betakernel}
\title{Title}
\usage{
betakernel(x, n = 512, lambda = 0.02, bcmethod = "beta1")
}
\arguments{
\item{x}{}

\item{n}{}

\item{lambda}{}

\item{bcmethod}{}
}
\value{
a list of x and y values. The y values describe the expected frequency of observations at value x.
}
\description{
Title
}

